Refactor jira-client.ts into modular structure
Split the monolithic 1072-line jira-client.ts file into smaller, focused modules:

- source/jira/types.ts: All TypeScript type definitions
- source/jira/utils.ts: Utility functions for configuration and data processing
- source/jira/client.ts: JiraClient class with HTTP operations
- source/jira-client.ts: Now acts as a re-export facade for backward compatibility

This refactoring improves code maintainability and follows the single responsibility principle. All existing functionality is preserved and tests pass.
